# openapi-gestion.yaml
...
openapi: 3.0.3
info:
  title: PID Gijón — API de gestión
  version: 1.0.0
  description: 'API de lectura y gestión de la Plataforma de Integración de Datos del Ayuntamiento de
    Gijón.


    El dato **no entra por aquí**: la ingesta va por el núcleo FIWARE (sensor → IoT Agent → Orion-LD →
    RabbitMQ → consumidores). Esta API cubre entidades, dispositivos, series temporales, cuadros de mando,
    alarmas y preferencias.


    Autenticación por *bearer token* JWT emitido por Keycloak. La API no participa en la sesión: el
    token se obtiene y se renueva contra el proveedor de identidad (código de autorización + PKCE para
    clientes de navegador). Las rutas bajo `/internal` no llevan token: solo son accesibles desde la
    red interna y las consume el gateway APISIX para validar permiso de escritura por tenant.'
  license:
    name: EUPL-1.2
    url: https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
servers:
- url: /api
  description: Ruta base del servicio
security:
- bearerAuth: []
...
paths:
...
  /V1/groups/{group}:
    delete:
      tags:
      - Grupos de entidades
      summary: Remove the specified resource from storage.
      operationId: EntityGroup_destroy_delete
      parameters:
      - name: group
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Eliminado
        '401':
          description: No autenticado
        '403':
          description: Sin permiso
        '404':
          description: No encontrado
    get:
      tags:
      - Grupos de entidades
      summary: Display the specified resource.
      operationId: EntityGroup_show_get
      parameters:
      - name: group
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
        '401':
          description: No autenticado
        '403':
          description: Sin permiso
        '404':
          description: No encontrado
    put:
      tags:
      - Grupos de entidades
      summary: Update the specified resource in storage.
      operationId: EntityGroup_update_put
      parameters:
      - name: group
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
        '401':
          description: No autenticado
        '403':
          description: Sin permiso
        '404':
          description: No encontrado
    patch:
      tags:
      - Grupos de entidades
      summary: Update the specified resource in storage.
      operationId: EntityGroup_update_patch
      parameters:
      - name: group
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
        '401':
          description: No autenticado
        '403':
          description: Sin permiso
        '404':
          description: No encontrado
  /V1/groups/{id}/status:
    patch:
      tags:
      - Grupos de entidades
      summary: Change an AssetIntervention's status and cascade it to all its member incidents (their
        status
      operationId: EntityGroup_updateStatus_patch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
        '401':
          description: No autenticado
        '403':
          description: Sin permiso
        '404':
          description: No encontrado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido por Keycloak (flujo de código con PKCE). Se obtiene y se renueva contra el proveedor de identidad.